+++ packages/twenty-server/src/engine/metadata-modules/remote-server/remote-table/distant-table/types/distant-table.ts
@@ -1,5 +1,5 @@
 import { PostgresTableSchemaColumn } from 'src/engine/metadata-modules/remote-server/types/postgres-table-schema-column';
 
 export type DistantTables = {
-  [tableName: string]: PostgresTableSchemaColumn[];
+  [distantTableName: string]: PostgresTableSchemaColumn[];
 };
