+++ packages/twenty-server/src/engine/metadata-modules/remote-server/remote-table/remote-table.resolver.ts
@@ -43,11 +43,11 @@ export class RemoteTableResolver {
   }
 
   @Mutation(() => RemoteTableDTO)
-  async updateRemoteTabletoDistantTable(
+  async syncRemoteTableSchemaChanges(
     @Args('input') input: RemoteTableInput,
     @AuthWorkspace() { id: workspaceId }: Workspace,
   ) {
-    return this.remoteTableService.updateRemoteTableToDistantTable(
+    return this.remoteTableService.syncRemoteTableSchemaChanges(
       input,
       workspaceId,
     );
