+++ packages/twenty-server/src/engine/metadata-modules/remote-server/remote-table/remote-table.service.ts
@@ -9,10 +9,7 @@ import {
   RemoteServerType,
   RemoteServerEntity,
 } from 'src/engine/metadata-modules/remote-server/remote-server.entity';
-import {
-  RemoteTableStatus,
-  DistantTableUpdate,
-} from 'src/engine/metadata-modules/remote-server/remote-table/dtos/remote-table.dto';
+import { RemoteTableStatus } from 'src/engine/metadata-modules/remote-server/remote-table/dtos/remote-table.dto';
 import {
   mapUdtNameToFieldType,
   mapUdtNameToFieldSettings,
@@ -30,17 +27,10 @@ import { WorkspaceDataSourceService } from 'src/engine/workspace-datasource/work
 import { RemoteTableEntity } from 'src/engine/metadata-modules/remote-server/remote-table/remote-table.entity';
 import { getRemoteTableLocalName } from 'src/engine/metadata-modules/remote-server/remote-table/utils/get-remote-table-local-name.util';
 import { DistantTableService } from 'src/engine/metadata-modules/remote-server/remote-table/distant-table/distant-table.service';
-import { DistantTables } from 'src/engine/metadata-modules/remote-server/remote-table/distant-table/types/distant-table';
-import { getForeignTableColumnName as convertToForeignTableColumnName } from 'src/engine/metadata-modules/remote-server/remote-table/foreign-table/utils/get-foreign-table-column-name.util';
 import { PostgresTableSchemaColumn } from 'src/engine/metadata-modules/remote-server/types/postgres-table-schema-column';
 import { fetchTableColumns } from 'src/engine/metadata-modules/remote-server/remote-table/utils/fetch-table-columns.util';
 import { ForeignTableService } from 'src/engine/metadata-modules/remote-server/remote-table/foreign-table/foreign-table.service';
-import {
-  WorkspaceMigrationAlterForeignTableAlteration,
-  WorkspaceMigrationColumnActionType,
-  WorkspaceMigrationColumnCreate,
-  WorkspaceMigrationColumnDrop,
-} from 'src/engine/metadata-modules/workspace-migration/workspace-migration.entity';
+import { RemoteTableSchemaUpdateService } from 'src/engine/metadata-modules/remote-server/remote-table/remote-table-schema-update/remote-table-schema-update.service';
 
 export class RemoteTableService {
   private readonly logger = new Logger(RemoteTableService.name);
@@ -59,6 +49,7 @@ export class RemoteTableService {
     private readonly distantTableService: DistantTableService,
     private readonly foreignTableService: ForeignTableService,
     private readonly workspaceDataSourceService: WorkspaceDataSourceService,
+    private readonly remoteTableSchemaUpdateService: RemoteTableSchemaUpdateService,
   ) {}
 
   public async findDistantTablesWithStatusByServerId(
@@ -105,7 +96,7 @@ export class RemoteTableService {
       return distantTablesWithStatus;
     }
 
-    return this.getDistantTablesWithUpdates({
+    return this.remoteTableSchemaUpdateService.getDistantTablesWithUpdates({
       remoteServerSchema: remoteServer.schema,
       workspaceId,
       remoteTables: currentRemoteTables,
@@ -275,7 +266,7 @@ export class RemoteTableService {
     }
   }
 
-  public async updateRemoteTableToDistantTable(
+  public async syncRemoteTableSchemaChanges(
     input: RemoteTableInput,
     workspaceId: string,
   ) {
@@ -321,12 +312,13 @@ export class RemoteTableService {
       remoteTable.localTableName,
     );
 
-    const alterations = this.computeForeignTableAlterations(
-      foreignTableColumns,
-      distantTableColumns,
-    );
+    const columnsUpdates =
+      this.remoteTableSchemaUpdateService.computeForeignTableColumnsUpdates(
+        foreignTableColumns,
+        distantTableColumns,
+      );
 
-    if (isEmpty(alterations)) {
+    if (isEmpty(columnsUpdates)) {
       this.logger.log(
         `No update to perform on table "${remoteTable.localTableName}" for workspace ${workspaceId}`,
       );
@@ -341,7 +333,7 @@ export class RemoteTableService {
     const updatedTable = await this.foreignTableService.updateForeignTable(
       remoteTable.localTableName,
       workspaceId,
-      alterations,
+      columnsUpdates,
     );
 
     return updatedTable;
@@ -446,155 +438,4 @@ export class RemoteTableService {
       }
     }
   }
-
-  private async getDistantTablesWithUpdates({
-    remoteServerSchema,
-    workspaceId,
-    remoteTables,
-    distantTables,
-  }: {
-    remoteServerSchema: string;
-    workspaceId: string;
-    remoteTables: RemoteTableEntity[];
-    distantTables: DistantTables;
-  }) {
-    const schemaPendingUpdates =
-      await this.getSchemaUpdatesBetweenForeignAndDistantTables({
-        workspaceId,
-        remoteTables,
-        distantTables,
-      });
-
-    const remoteTablesDistantNames = remoteTables.map(
-      (remoteTable) => remoteTable.distantTableName,
-    );
-
-    const distantTablesWithUpdates = Object.keys(distantTables).map(
-      (tableName) => ({
-        name: tableName,
-        schema: remoteServerSchema,
-        status: remoteTablesDistantNames.includes(tableName)
-          ? RemoteTableStatus.SYNCED
-          : RemoteTableStatus.NOT_SYNCED,
-        schemaPendingUpdates: schemaPendingUpdates[tableName],
-      }),
-    );
-
-    const deletedTables = Object.entries(schemaPendingUpdates)
-      .filter(([_tableName, updates]) =>
-        updates.includes(DistantTableUpdate.TABLE_DELETED),
-      )
-      .map(([tableName, updates]) => ({
-        name: tableName,
-        schema: remoteServerSchema,
-        status: RemoteTableStatus.SYNCED,
-        schemaPendingUpdates: updates,
-      }));
-
-    return distantTablesWithUpdates.concat(deletedTables);
-  }
-
-  private async getSchemaUpdatesBetweenForeignAndDistantTables({
-    workspaceId,
-    remoteTables,
-    distantTables,
-  }: {
-    workspaceId: string;
-    remoteTables: RemoteTableEntity[];
-    distantTables: DistantTables;
-  }): Promise<{ [tablename: string]: DistantTableUpdate[] }> {
-    const updates = {};
-
-    for (const remoteTable of remoteTables) {
-      const distantTable = distantTables[remoteTable.distantTableName];
-      const tableName = remoteTable.distantTableName;
-
-      if (!distantTable) {
-        updates[tableName] = [DistantTableUpdate.TABLE_DELETED];
-        continue;
-      }
-
-      const foreignTable = await fetchTableColumns(
-        this.workspaceDataSourceService,
-        workspaceId,
-        remoteTable.localTableName,
-      );
-
-      const { columnsAdded, columnsDeleted } = this.compareForeignTableColumns(
-        foreignTable,
-        distantTable,
-      );
-
-      if (columnsAdded.length > 0) {
-        updates[tableName] = [
-          ...(updates[tableName] || []),
-          DistantTableUpdate.COLUMNS_ADDED,
-        ];
-      }
-      if (columnsDeleted.length > 0) {
-        updates[tableName] = [
-          ...(updates[tableName] || []),
-          DistantTableUpdate.COLUMNS_DELETED,
-        ];
-      }
-    }
-
-    return updates;
-  }
-
-  private compareForeignTableColumns = (
-    foreignTableColumns: PostgresTableSchemaColumn[],
-    distantTableColumns: PostgresTableSchemaColumn[],
-  ) => {
-    const foreignTableColumnNames = foreignTableColumns.map(
-      (column) => column.columnName,
-    );
-    const distantTableColumnsWithConvertedName = distantTableColumns.map(
-      (column) => {
-        return {
-          name: convertToForeignTableColumnName(column.columnName),
-          type: column.dataType,
-        };
-      },
-    );
-
-    const columnsAdded = distantTableColumnsWithConvertedName.filter(
-      (column) => !foreignTableColumnNames.includes(column.name),
-    );
-    const columnsDeleted = foreignTableColumnNames.filter(
-      (columnName) =>
-        !distantTableColumnsWithConvertedName
-          .map((column) => column.name)
-          .includes(columnName),
-    );
-
-    return {
-      columnsAdded,
-      columnsDeleted,
-    };
-  };
-
-  private computeForeignTableAlterations = (
-    foreignTableColumns: PostgresTableSchemaColumn[],
-    distantTableColumns: PostgresTableSchemaColumn[],
-  ): WorkspaceMigrationAlterForeignTableAlteration[] => {
-    const { columnsAdded, columnsDeleted } = this.compareForeignTableColumns(
-      foreignTableColumns,
-      distantTableColumns,
-    );
-    const columnsAddedAlterations: WorkspaceMigrationColumnCreate[] =
-      columnsAdded.map((columnAdded) => ({
-        action: WorkspaceMigrationColumnActionType.CREATE,
-        columnName: columnAdded.name,
-        columnType: columnAdded.type,
-      }));
-
-    const columnsDeletedAlterations: WorkspaceMigrationColumnDrop[] =
-      columnsDeleted.map((columnDeleted) => ({
-        action: WorkspaceMigrationColumnActionType.DROP,
-        columnName: columnDeleted,
-      }));
-
-    return [...columnsAddedAlterations, ...columnsDeletedAlterations];
-  };
 }
